# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Min Khaung Tea House & Restaurant Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Table {
  id: number;
  tableNumber: string;
  /** hall or aircon */
  zone: string;
  capacity: number;
  /** available, occupied, payment_pending, dirty */
  status: string;
  qrCode?: string;
  posX: number;
  posY: number;
  /** @nullable */
  currentOrderId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateTableBody {
  tableNumber: string;
  zone: string;
  capacity: number;
  posX: number;
  posY: number;
}

export interface UpdateTableBody {
  tableNumber?: string;
  zone?: string;
  capacity?: number;
  status?: string;
  posX?: number;
  posY?: number;
  /** @nullable */
  currentOrderId?: number | null;
}

export interface MenuCategory {
  id: number;
  name: string;
  nameMyanmar: string;
  sortOrder: number;
  createdAt: string;
}

export interface CreateMenuCategoryBody {
  name: string;
  nameMyanmar: string;
  sortOrder?: number;
}

export interface UpdateMenuCategoryBody {
  name?: string;
  nameMyanmar?: string;
  sortOrder?: number;
}

export interface MenuItem {
  id: number;
  categoryId: number;
  name: string;
  nameMyanmar: string;
  /** @nullable */
  description?: string | null;
  price: string;
  /** @nullable */
  imageUrl?: string | null;
  available: string;
  /**
   * JSON string of customization options
   * @nullable
   */
  customizationOptions?: string | null;
  /** @nullable */
  qrCode?: string | null;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateMenuItemBody {
  categoryId: number;
  name: string;
  nameMyanmar: string;
  description?: string;
  price: string;
  imageUrl?: string;
  available?: string;
  customizationOptions?: string;
  sortOrder?: number;
}

export interface UpdateMenuItemBody {
  categoryId?: number;
  name?: string;
  nameMyanmar?: string;
  description?: string;
  price?: string;
  imageUrl?: string;
  available?: string;
  customizationOptions?: string;
  sortOrder?: number;
}

export interface Order {
  id: number;
  tableId: number;
  tableNumber: string;
  /** open, ready_to_pay, paid, cancelled */
  status: string;
  subtotal: string;
  airconFee: string;
  taxAmount: string;
  totalAmount: string;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  staffId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface OrderItem {
  id: number;
  orderId: number;
  menuItemId: number;
  menuItemName: string;
  quantity: number;
  unitPrice: string;
  /** @nullable */
  customizations?: string | null;
  /** new, cooking, ready, served */
  kitchenStatus: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface OrderDetail {
  id: number;
  tableId: number;
  tableNumber: string;
  status: string;
  subtotal: string;
  airconFee: string;
  taxAmount: string;
  totalAmount: string;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  staffId?: number | null;
  items: OrderItem[];
  createdAt: string;
  updatedAt: string;
}

export interface OrderItemInput {
  menuItemId: number;
  quantity: number;
  customizations?: string;
  notes?: string;
}

export interface CreateOrderBody {
  tableId: number;
  notes?: string;
  staffId?: number;
  items?: OrderItemInput[];
}

export interface UpdateOrderBody {
  status?: string;
  notes?: string;
  paymentMethod?: string;
}

export interface AddOrderItemBody {
  menuItemId: number;
  quantity: number;
  customizations?: string;
  notes?: string;
}

export interface UpdateOrderItemBody {
  quantity?: number;
  kitchenStatus?: string;
  customizations?: string;
  notes?: string;
}

export interface KitchenOrder {
  orderId: number;
  tableNumber: string;
  zone: string;
  orderTime: string;
  items: OrderItem[];
}

export interface UpdateKitchenItemStatusBody {
  /** new, cooking, ready, served */
  kitchenStatus: string;
}

export interface Payment {
  id: number;
  orderId: number;
  tableNumber: string;
  amount: string;
  /** cash, wave_pay, kbz_pay, aya_pay, cb_pay */
  paymentMethod: string;
  /** pending, completed, refunded */
  status: string;
  /** @nullable */
  cashierId?: number | null;
  receiptNumber: string;
  createdAt: string;
}

export interface CreatePaymentBody {
  orderId: number;
  paymentMethod: string;
  cashierId?: number;
}

export interface InventoryItem {
  id: number;
  name: string;
  unit: string;
  currentStock: string;
  minimumStock: string;
  cost: string;
  /** @nullable */
  supplierId?: number | null;
  /** @nullable */
  lastRestockedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateInventoryItemBody {
  name: string;
  unit: string;
  currentStock: string;
  minimumStock: string;
  cost: string;
}

export interface UpdateInventoryItemBody {
  name?: string;
  unit?: string;
  currentStock?: string;
  minimumStock?: string;
  cost?: string;
}

export interface StaffMember {
  id: number;
  name: string;
  /** admin, cashier, waitress, kitchen */
  role: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  pin?: string | null;
  active: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateStaffMemberBody {
  name: string;
  role: string;
  phone?: string;
  email?: string;
  pin?: string;
}

export interface UpdateStaffMemberBody {
  name?: string;
  role?: string;
  phone?: string;
  email?: string;
  pin?: string;
  active?: string;
}

export interface Transaction {
  id: number;
  /** deposit, withdrawal, expense */
  type: string;
  amount: string;
  description: string;
  /** @nullable */
  staffId?: number | null;
  /** @nullable */
  referenceNumber?: string | null;
  createdAt: string;
}

export interface CreateTransactionBody {
  type: string;
  amount: string;
  description: string;
  staffId?: number;
  referenceNumber?: string;
}

export interface Settings {
  id: number;
  restaurantName: string;
  taxRate: string;
  airconFee: string;
  currency: string;
  /** @nullable */
  receiptFooter?: string | null;
  updatedAt: string;
}

export interface UpdateSettingsBody {
  restaurantName?: string;
  taxRate?: string;
  airconFee?: string;
  currency?: string;
  receiptFooter?: string;
}

export interface DashboardSummary {
  todaySales: string;
  todayOrders: number;
  activeOrders: number;
  availableTables: number;
  occupiedTables: number;
  pendingPaymentTables: number;
  lowStockItems: number;
  monthSales: string;
}

export interface SalesChartPoint {
  date: string;
  sales: string;
  orders: number;
}

export interface TopItem {
  menuItemId: number;
  name: string;
  totalSold: number;
  totalRevenue: string;
}

export type ListMenuItemsParams = {
  /**
   * @nullable
   */
  categoryId?: number | null;
  /**
   * @nullable
   */
  available?: string | null;
};

export type ListOrdersParams = {
  /**
   * @nullable
   */
  status?: string | null;
  /**
   * @nullable
   */
  tableId?: number | null;
  /**
   * @nullable
   */
  date?: string | null;
};

export type ListPaymentsParams = {
  /**
   * @nullable
   */
  date?: string | null;
  /**
   * @nullable
   */
  status?: string | null;
};

export type ListTransactionsParams = {
  /**
   * @nullable
   */
  type?: string | null;
  /**
   * @nullable
   */
  startDate?: string | null;
  /**
   * @nullable
   */
  endDate?: string | null;
};