# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Min Khaung Tea House & Restaurant Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /menu-categories:
    get:
      operationId: listMenuCategories
      tags: [menu]
      summary: List menu categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuCategory"
    post:
      operationId: createMenuCategory
      tags: [menu]
      summary: Create a menu category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMenuCategoryBody"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuCategory"
  /menu-categories/{id}:
    patch:
      operationId: updateMenuCategory
      tags: [menu]
      summary: Update a menu category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMenuCategoryBody"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuCategory"
    delete:
      operationId: deleteMenuCategory
      tags: [menu]
      summary: Delete a menu category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /menu-items:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List all menu items
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: available
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createMenuItem
      tags: [menu]
      summary: Create a menu item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMenuItemBody"
      responses:
        "201":
          description: Created menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
  /menu-items/{id}:
    get:
      operationId: getMenuItem
      tags: [menu]
      summary: Get a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found
    patch:
      operationId: updateMenuItem
      tags: [menu]
      summary: Update a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMenuItemBody"
      responses:
        "200":
          description: Updated menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    delete:
      operationId: deleteMenuItem
      tags: [menu]
      summary: Delete a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── ORDERS ────────────────────────────────────────────────────────────────
...
  /kitchen/items/{itemId}/status:
    patch:
      operationId: updateKitchenItemStatus
      tags: [kitchen]
      summary: Update kitchen item cooking status
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateKitchenItemStatusBody"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderItem"

  # ─── PAYMENTS ──────────────────────────────────────────────────────────────
...
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List inventory items
      responses:
        "200":
          description: Inventory list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Add inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryItemBody"
      responses:
        "201":
          description: Created inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/{id}:
    patch:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInventoryItemBody"
      responses:
        "200":
          description: Updated inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Delete inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── STAFF ─────────────────────────────────────────────────────────────────
...
  /finance/transactions:
    get:
      operationId: listTransactions
      tags: [finance]
      summary: List financial transactions
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: ["string", "null"]
        - name: startDate
          in: query
          required: false
          schema:
            type: ["string", "null"]
        - name: endDate
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Transactions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [finance]
      summary: Create a financial transaction (bank deposit/withdrawal)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionBody"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"

  # ─── SETTINGS ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MenuCategory:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameMyanmar:
          type: string
        sortOrder:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - nameMyanmar
        - sortOrder
        - createdAt
    CreateMenuCategoryBody:
      type: object
      properties:
        name:
          type: string
        nameMyanmar:
          type: string
        sortOrder:
          type: integer
      required:
        - name
        - nameMyanmar
    UpdateMenuCategoryBody:
      type: object
      properties:
        name:
          type: string
        nameMyanmar:
          type: string
        sortOrder:
          type: integer
    MenuItem:
      type: object
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        name:
          type: string
        nameMyanmar:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: string
        imageUrl:
          type: ["string", "null"]
        available:
          type: string
        customizationOptions:
          type: ["string", "null"]
          description: JSON string of customization options
        qrCode:
          type: ["string", "null"]
        sortOrder:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - categoryId
        - name
        - nameMyanmar
        - price
        - available
        - sortOrder
        - createdAt
        - updatedAt
    CreateMenuItemBody:
      type: object
      properties:
        categoryId:
          type: integer
        name:
          type: string
        nameMyanmar:
          type: string
        description:
          type: string
        price:
          type: string
        imageUrl:
          type: string
        available:
          type: string
        customizationOptions:
          type: string
        sortOrder:
          type: integer
      required:
        - categoryId
        - name
        - nameMyanmar
        - price
    UpdateMenuItemBody:
      type: object
      properties:
        categoryId:
          type: integer
        name:
          type: string
        nameMyanmar:
          type: string
        description:
          type: string
        price:
          type: string
        imageUrl:
          type: string
        available:
          type: string
        customizationOptions:
          type: string
        sortOrder:
          type: integer

    # ─── ORDER SCHEMAS ────────────────────────────────────────────────────
...
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        menuItemId:
          type: integer
        menuItemName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: string
        customizations:
          type: ["string", "null"]
        kitchenStatus:
          type: string
          description: "new, cooking, ready, served"
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - orderId
        - menuItemId
        - menuItemName
        - quantity
        - unitPrice
        - kitchenStatus
        - createdAt
...
    UpdateKitchenItemStatusBody:
      type: object
      properties:
        kitchenStatus:
          type: string
          description: "new, cooking, ready, served"
      required:
        - kitchenStatus

    # ─── PAYMENT SCHEMAS ──────────────────────────────────────────────────
...
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        unit:
          type: string
        currentStock:
          type: string
        minimumStock:
          type: string
        cost:
          type: string
        supplierId:
          type: ["integer", "null"]
        lastRestockedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - unit
        - currentStock
        - minimumStock
        - cost
        - createdAt
        - updatedAt
    CreateInventoryItemBody:
      type: object
      properties:
        name:
          type: string
        unit:
          type: string
        currentStock:
          type: string
        minimumStock:
          type: string
        cost:
          type: string
      required:
        - name
        - unit
        - currentStock
        - minimumStock
        - cost
    UpdateInventoryItemBody:
      type: object
      properties:
        name:
          type: string
        unit:
          type: string
        currentStock:
          type: string
        minimumStock:
          type: string
        cost:
          type: string

    # ─── STAFF SCHEMAS ────────────────────────────────────────────────────
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          description: "deposit, withdrawal, expense"
        amount:
          type: string
        description:
          type: string
        staffId:
          type: ["integer", "null"]
        referenceNumber:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - type
        - amount
        - description
        - createdAt
    CreateTransactionBody:
      type: object
      properties:
        type:
          type: string
        amount:
          type: string
        description:
          type: string
        staffId:
          type: integer
        referenceNumber:
          type: string
      required:
        - type
        - amount
        - description

    # ─── SETTINGS SCHEMAS ─────────────────────────────────────────────────
...