# lib/api-zod/src/generated/types/createMenuCategoryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Min Khaung Tea House & Restaurant Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMenuCategoryBody {
  name: string;
  nameMyanmar: string;
  sortOrder?: number;
}