# lib/api-zod/src/generated/types/createMenuItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Min Khaung Tea House & Restaurant Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMenuItemBody {
  categoryId: number;
  name: string;
  nameMyanmar: string;
  description?: string;
  price: string;
  imageUrl?: string;
  available?: string;
  customizationOptions?: string;
  sortOrder?: number;
}